Switch ROW_FORMAT to COMPRESSED
I'm not sure what the intention is but docs [1] says that
ROW_FORMAT=DYNAMIC and specifying a KEY_BLOCK_SIZE is an invalid
combination:

  KEY_BLOCK_SIZE is specified with REDUNDANT, COMPACT or DYNAMIC
  row format -> KEY_BLOCK_SIZE is ignored

Looks like you would need innodb_strict_mode off to allow this
setup to be run without a hard error. Not sure how/why the db is
able to export this config.

I might be missing something...

[1] https://dev.mysql.com/doc/refman/5.7/en/innodb-compression-syntax-warnings.html